o Support a compile-time -DNO_FW_PUNCH for portability   (and those of us that don't want the functionality). o Don't assume sizeof(long) == 4. Ok'd by: Charles Mott <cmott@srv.net>
git-svn-id: svn+ssh://svn.freebsd.org/base/head@35314 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netinet/libalias/alias_irc.c
@@ -229,7 +229,7 @@ lFOUND_CTCP:
 				 alias_address = GetAliasAddress(link);
 				 iCopy += snprintf(&newpacket[iCopy],
 										 sizeof(newpacket)-iCopy, 
-										 "%lu ", htonl(alias_address.s_addr));
+										 "%lu ", (u_long)htonl(alias_address.s_addr));
 				 if( iCopy >= sizeof(newpacket) ) { /* Truncated/fit exactly - bad news */
 					 DBprintf(("DCC constructed packet overflow.\n"));
 					 goto lBAD_CTCP;
